requirements.txt
alembic==1.7.5
aniso8601==9.0.1
astroid==2.9.1
attrs==21.4.0
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cors==1.0.1
cryptography==36.0.1
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.7.1
loguru==0.5.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
mccabe==0.6.1
moto==2.3.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.11.1
PyJWT==2.3.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
requests-file==1.5.1
responses==0.16.0
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.29
tldextract==3.1.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.4.0
